+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 2.3.6
 * Added JSON helper methods to PieceConfig
 * Item property in ItemConfig is publicly readable now so it can be serialized
+* Added slider for numerical values in the mod settings GUI
 
 ## Version 2.3.5
 * Added support for custom Obliterator/Incinerator item conversions