+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added slider for numerical values in the mod settings GUI
 * Fixed button config in mod settings for control keys
 * Added support for BepInEx' KeyboardShortcuts in ButtonConfig and InputManager
+* Fixed the localization for Jötunn tokens
 
 ## Version 2.3.5
 * Added support for custom Obliterator/Incinerator item conversions